requirements.txt
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.1.4
pathway==0.7.2
python-dotenv==1.0.0
requests==2.31.0
sentence-transformers==2.2.2
fastapi==0.109.2
uvicorn==0.27.1
websockets==12.0